TEngineHostCounters in separate file
