+++ ydb/core/engine/minikql/minikql_engine_host.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "minikql_engine_host_counters.h"
 #include "change_collector_iface.h"
 
 #include <util/generic/cast.h>
@@ -11,56 +12,6 @@
 
 namespace NKikimr {
 namespace NMiniKQL {
-
-struct TEngineHostCounters {
-    ui64 NSelectRow = 0;
-    ui64 NSelectRange = 0;
-    ui64 NUpdateRow  = 0;
-    ui64 NEraseRow = 0;
-
-    ui64 SelectRowRows = 0;
-    ui64 SelectRowBytes = 0;
-    ui64 SelectRangeRows = 0;
-    ui64 SelectRangeBytes = 0;
-    ui64 SelectRangeDeletedRowSkips = 0;
-    ui64 UpdateRowBytes = 0;
-    ui64 EraseRowBytes = 0;
-
-    ui64 InvisibleRowSkips = 0;
-
-    TEngineHostCounters& operator+=(const TEngineHostCounters& other) {
-        NSelectRow += other.NSelectRow;
-        NSelectRange += other.NSelectRange;
-        NUpdateRow += other.NUpdateRow;
-        NEraseRow += other.NEraseRow;
-        SelectRowRows += other.SelectRowRows;
-        SelectRowBytes += other.SelectRowBytes;
-        SelectRangeRows += other.SelectRangeRows;
-        SelectRangeBytes += other.SelectRangeBytes;
-        SelectRangeDeletedRowSkips += other.SelectRangeDeletedRowSkips;
-        UpdateRowBytes += other.UpdateRowBytes;
-        EraseRowBytes += other.EraseRowBytes;
-        InvisibleRowSkips += other.InvisibleRowSkips;
-        return *this;
-    }
-
-    TString ToString() const {
-        return TStringBuilder()
-            << "{NSelectRow: " << NSelectRow
-            << ", NSelectRange: " << NSelectRange
-            << ", NUpdateRow: " << NUpdateRow
-            << ", NEraseRow: " << NEraseRow
-            << ", SelectRowRows: " << SelectRowRows
-            << ", SelectRowBytes: " << SelectRowBytes
-            << ", SelectRangeRows: " << SelectRangeRows
-            << ", SelectRangeBytes: " << SelectRangeBytes
-            << ", UpdateRowBytes: " << UpdateRowBytes
-            << ", EraseRowBytes: " << EraseRowBytes
-            << ", SelectRangeDeletedRowSkips: " << SelectRangeDeletedRowSkips
-            << ", InvisibleRowSkips: " << InvisibleRowSkips
-            << "}";
-    }
-};
 
 struct IKeyAccessSampler : public TThrRefBase {
     using TPtr = TIntrusivePtr<IKeyAccessSampler>;
